+++ src/commands/General/invite.ts
@@ -5,6 +5,7 @@ export default class extends Command {
     public constructor(store: CommandStore, file: string[], directory: string) {
         super(store, file, directory, {
             guarded: true,
+            runIn: ['text', 'dm'],
             description: language => language.get('COMMAND_INVITE_DESCRIPTION')
         });
     }