+++ src/commands/General/ping.ts
@@ -5,6 +5,7 @@ export default class extends Command {
     constructor(store: CommandStore, file: string[], directory: string) {
         super(store, file, directory, {
             guarded: true,
+            runIn: ['text', 'dm'],
             description: language => language.get('COMMAND_PING_DESCRIPTION')
         });
     }