+++ src/lib/structures/permissionLevel.ts
@@ -1,11 +1,11 @@
-import { PermissionLevels, KlasaUser, KlasaMessage } from 'klasa';
+import { PermissionLevels, KlasaMessage } from 'klasa';
 import { GuildMember, Guild } from 'discord.js';
 
 export default new PermissionLevels()
 // @everyone
     .add(0, () => true)
 // PLAYER
-    .add(2, ({ author }) => Boolean((author as KlasaUser).settings.get('inventory.id')))
+// .add(2, ({ author }) => Boolean((author as KlasaUser).settings.get('inventory.id')))
 // MANAGE_MESSAGES
     .add(4, ({ guild, member }: KlasaMessage) => guild ? true : false && (member as GuildMember).permissions.has('MANAGE_MESSAGES'), { fetch: true })
 // ADMINISTRATOR